Move tests out of index.rs
